backend/tests/test_auth.py
import pytest
from fastapi.testclient import TestClient
from main import app
from settings import settings
import jwt

def test_jwt_verification_valid_token():
    """Test that a valid JWT token is properly verified."""
    # Create a valid token for testing
    test_user_id = "test_user_123"
    token = jwt.encode(
        {"user_id": test_user_id, "exp": 9999999999},  # Far future expiration
        settings.better_auth_secret,
        algorithm="HS256"
    )

    with TestClient(app) as client:
        # This test would require a route that uses the auth dependency
        # For now, we'll just verify the token creation and decoding works
        decoded = jwt.decode(
            token,
            settings.better_auth_secret,
            algorithms=["HS256"]
        )
        assert decoded["user_id"] == test_user_id

def test_jwt_verification_invalid_token():
    """Test that an invalid JWT token raises appropriate error."""
    # Create an invalid token by using wrong secret
    test_user_id = "test_user_123"
    wrong_secret = "wrong_secret"
    token = jwt.encode(
        {"user_id": test_user_id, "exp": 9999999999},  # Far future expiration
        wrong_secret,
        algorithm="HS256"
    )

    with TestClient(app) as client:
        # This test would require a route that uses the auth dependency
        # For now, we'll just verify that decoding with wrong secret fails
        with pytest.raises(jwt.InvalidSignatureError):
            jwt.decode(
                token,
                settings.better_auth_secret,
                algorithms=["HS256"]
            )

def test_jwt_verification_expired_token():
    """Test that an expired JWT token raises appropriate error."""
    # Create an expired token
    test_user_id = "test_user_123"
    token = jwt.encode(
        {"user_id": test_user_id, "exp": 1},  # Expired long ago
        settings.better_auth_secret,
        algorithm="HS256"
    )

    with TestClient(app) as client:
        # This test would require a route that uses the auth dependency
        # For now, we'll just verify that decoding an expired token fails
        with pytest.raises(jwt.ExpiredSignatureError):
            jwt.decode(
                token,
                settings.better_auth_secret,
                algorithms=["HS256"]
            )