backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
pydantic-settings
pyjwt
passlib[bcrypt]
python-multipart
python-jose[cryptography]
cryptography
httpx
requests
python-dotenv
pytest
pytest-asyncio
pytest-cov==5.0.0
psycopg2-binary
openai
async-openai